Move logging messages to network_messages dir
